+++ TeamCode/src/main/java/org/nknsd/robotics/team/ExtensionMonkeyProgram.java
@@ -2,11 +2,8 @@ package org.nknsd.robotics.team;
 
 import org.nknsd.robotics.framework.NKNComponent;
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.components.ArmRotator;
 import org.nknsd.robotics.team.components.ExtensionHandler;
-import org.nknsd.robotics.team.components.PotentiometerHandler;
 import org.nknsd.robotics.team.components.testfiles.ExtensionMonkey;
-import org.nknsd.robotics.team.components.testfiles.RotatorMonkey;
 
 import java.util.List;
