+++ TeamCode/src/main/java/org/nknsd/robotics/team/RotatorMonkeyProgram.java
@@ -2,12 +2,9 @@ package org.nknsd.robotics.team;
 
 import org.nknsd.robotics.framework.NKNComponent;
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.components.ArmRotator;
-import org.nknsd.robotics.team.components.GamePadHandler;
+import org.nknsd.robotics.team.components.RotationHandler;
 import org.nknsd.robotics.team.components.PotentiometerHandler;
-import org.nknsd.robotics.team.components.WheelDriver;
-import org.nknsd.robotics.team.components.WheelHandler;
-import org.nknsd.robotics.team.components.testfiles.RotatorMonkey;
+import org.nknsd.robotics.team.components.testfiles.RotationMonkey;
 
 import java.util.List;
 
@@ -18,14 +15,14 @@ public class RotatorMonkeyProgram extends NKNProgram {
         components.add(potentiometerHandler);
         telemetryEnabled.add(potentiometerHandler);
 
-        ArmRotator armRotator = new ArmRotator("motorArmRotate", 0.05, 0.9);
-        armRotator.link(potentiometerHandler);
-        components.add(armRotator);
-        telemetryEnabled.add(armRotator);
+        RotationHandler rotationHandler = new RotationHandler("motorArmRotate", 0.05, 0.9);
+        rotationHandler.link(potentiometerHandler);
+        components.add(rotationHandler);
+        telemetryEnabled.add(rotationHandler);
 
-        RotatorMonkey rotatorMonkey = new RotatorMonkey();
-        rotatorMonkey.link(armRotator);
-        components.add(rotatorMonkey);
-        telemetryEnabled.add(rotatorMonkey);
+        RotationMonkey rotationMonkey = new RotationMonkey();
+        rotationMonkey.link(rotationHandler);
+        components.add(rotationMonkey);
+        telemetryEnabled.add(rotationMonkey);
     }
 }
