+++ TeamCode/src/main/java/org/nknsd/robotics/team/components/testfiles/ServoMonkey.java
@@ -6,7 +6,6 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.nknsd.robotics.framework.NKNComponent;
-import org.nknsd.robotics.team.components.ArmRotator;
 import org.nknsd.robotics.team.components.IntakeServoHandler;
 
 import java.util.concurrent.TimeUnit;
